Update_Table
Met with Luke McCormack and have re-read out the drought table. The new drought table includes all the plants on The Morton Arboretum website. This list was cross-referenced with the larger pool of data to get PSI TLP data where available for species.

The next steps will be to create Genera-level graphs for Acer, Quercus, Ulmus, and Magnolia.